Try to fix win64 buildbot.
packet-dbus.c(354) : error C2220: warning treated as error - no 'object' file generated
packet-dbus.c(354) : warning C4267: '=' : conversion from 'size_t' to 'int', possible loss of data

svn path=/trunk/; revision=43543

+++ epan/dissectors/packet-dbus.c
@@ -351,7 +351,7 @@ dissect_dbus_field_signature(tvbuff_t *tvb, dbus_info_t *dinfo, proto_tree *tree
 				offset = dissect_dbus_sig(tvb, dinfo, tree, offset, 's', &addr_val);
 				if (offset != -1)
 					SET_ADDRESS((field_code == DBUS_HEADER_FIELD_DESTINATION) ? &dinfo->pinfo->dst : &dinfo->pinfo->src,
-					            AT_STRINGZ, strlen(addr_val.str)+1, addr_val.str);
+					            AT_STRINGZ, (int)strlen(addr_val.str)+1, addr_val.str);
 				return offset;
 			}
 			break;
